splitting autoemailer into new file
